update data for weeks 2021-41 to 2021-50
There was an empty json file in the snapshots which broke the exporter.

Anyways. This repo is deprecated now. There is some other code waiting to take over
which exports json files with diffs between each snapshots. That saves **a lot** of
space. This repo is already 250mb in size and 1gb is the limit on github (without
the microsoft enterpiseness)